update Workout data structure - based on composite design pattern
test code in to read in data to new Workout data structure still WIP

may be using builder pattern in order to use Workout structure as basis for
creation of workouts. - TBD

A fair bit of the design / implementation of this is based on information / code in:
https://refactoring.guru/design-patterns/cpp

big thanks to the contributers to that site for providing me with basis for this app.